extrapolate the db logic into store.rs
